move main config to cmd package
